ceaning ex10 here and there

+++ applications/MPM_FEM/ex10/include/gmpm.hpp
@@ -35,10 +35,10 @@ class WindowFunction {
     double dw0;
     double dw1;
 
-    const static unsigned ak[5][5];
+    const static double ak[5][5];
 };
 
-const unsigned WindowFunction::ak[5][5] = {{1}, {1, 2}, {1, 3, 6}, {1, 4, 10, 20}, {1, 5, 15, 35, 70}};
+const double WindowFunction::ak[5][5] = {{1.}, {1., 2.}, {1., 3., 6.}, {1., 4., 10., 20.}, {1., 5., 15., 35., 70.}};
 
 void WindowFunction::BuildWeight (const std::vector <double> &Xv, const unsigned &pOrder, const double &x, const bool &nonLocal, const unsigned &Cn) {
 
@@ -48,54 +48,14 @@ void WindowFunction::BuildWeight (const std::vector <double> &Xv, const unsigned
   x1 = Xv[i];
 
   if (nonLocal) {
-
-
-//     double l0 = (x - x1) / (x0 - x1);
-//     double l1 = (x - x0) / (x1 - x0);
-//
-//     double l0p = 1. / (x0 - x1);
-//     double l1p = -l0p;
-//
-//     unsigned n = Cn + 1;
-//     double l0n = pow (l0, n);
-//     double l0nm1 = l0n / l0;
-//     double l1n = pow (l1, n);
-//     double l1nm1 = l1n / l1;
-//
-//     double Pl0 = l0nm1;
-//     double Pl1 = l1nm1;
-//     double dPl0 = 0.;
-//     double dPl1 = 0.;
-//     double l0nm1mk = l0nm1;
-//     double l1nm1mk = l1nm1;
-//     for (unsigned k = 1; k < n; k++) {
-//       l0nm1mk /= l0;
-//       l1nm1mk /= l1;
-//       Pl0 += l0nm1mk / boost::math::factorial<double> (k + 1);
-//       Pl1 += l1nm1mk / boost::math::factorial<double> (k + 1);
-//       dPl0 += (n - k) * l0nm1mk * l0p / boost::math::factorial<double> (k);
-//       dPl1 += (n - k) * l1nm1mk * l1p / boost::math::factorial<double> (k);
-//     }
-//
-//     w0 = Pl1 * boost::math::factorial<double> (n) * l0n;
-//     w1 = Pl0 * boost::math::factorial<double> (n) * l1n;
-//     dw0 = boost::math::factorial<double> (n) * l0nm1 * (n * l0p * Pl1 + l0 * dPl1);
-//     dw1 = boost::math::factorial<double> (n) * l1nm1 * (n * l1p * Pl0 + l1 * dPl0);
-
-
+      
     double l0 = (x - x1) / (x0 - x1);
     double l1 = (x - x0) / (x1 - x0);
 
     double dl0 = 1. / (x0 - x1);
     double dl1 = -dl0;
 
-
-
     unsigned n = Cn + 1;
-//     double l0n = pow (l0, n);
-//     double l0nm1 = l0n / l0;
-//     double l1n = pow (l1, n);
-//     double l1nm1 = l1n / l1;
 
     double Pl0 = ak[Cn][0];
     double Pl1 = ak[Cn][0];
